Fixed upgrade scripts for the switch_mod -> switch_model package renaming. * Squash bugs so tests run in clean environments * Modularize upgrade plugins to iterate over versions, allowing support for intermediate upgrades * Separating upgrade into sequential steps also simplifies individual upgrade scripts, lowering the barrier for making those type of edits * Added documentation for upgrade tests to ease the learning curve for developers
Other changes:
* Cleaned switch_mod.upgrade package to streamline future work and set clear patterns
* Moved optional imports in basic_exports.py to a function to avoid breaking tests for installations that don’t use that optional functionality
* Incremented example input versions to complete the directory upgrades
* Moved pandas into a required dependency because it is necessary for testing, and upgrading inputs is core functionality to switch (IMO)
