+++ examples/3zone_toy_stochastic_PySP/inputs/switch_inputs_version.txt
@@ -1,1 +1,1 @@
-2.0.0b1
+2.0.0b2
