+++ examples/carbon_cap/inputs/switch_inputs_version.txt
@@ -1,1 +1,1 @@
-2.0.0b1
+2.0.0b2
