+++ examples/production_cost_models/4plants_with_unserved_load/inputs/switch_inputs_version.txt
@@ -1,1 +1,1 @@
-2.0.0b1
+2.0.0b2
