+++ examples/production_cost_models/discrete_unit_commit/inputs/switch_inputs_version.txt
@@ -1,1 +1,1 @@
-2.0.0b1
+2.0.0b2
