+++ tests/upgrade_dat/custom_extension/inputs/modules.txt
@@ -1,4 +1,4 @@
-switch_model
+switch_mod
 project.no_commit
 fuel_cost
 sunk_costs
