+++ tests/upgrade_test.py
@@ -1,5 +1,17 @@
 # Copyright 2016 The Switch Authors. All rights reserved.
 # Licensed under the Apache License, Version 2, which is in the LICENSE file.
+"""
+Test upgrade functionality by:
+* Upgrading each example dataset found in tests/upgrade_dat
+* Solving the upgraded inputs
+* Validating that the objective function value is the same
+
+Note to developers: Do not update the input directories of tests/upgrade_dat
+or else this test won't actually test the upgrade functionality. See
+tests/upgrade_dat/README.txt for an explanation.
+
+"""
+
 
 import filecmp
 import os
@@ -12,8 +24,8 @@ import unittest
 import switch_model.solve
 import switch_model.utilities
 from switch_model.upgrade import upgrade_inputs
+import switch_model.upgrade.manager
 from .examples_test import get_expectation_path, read_file, write_file, TOP_DIR
-import switch_model.upgrade.upgrade_2_0_0b1 as upgrade_2_0_0b1
 
 UPDATE_EXPECTATIONS = False
 
@@ -40,9 +52,10 @@ def make_test(example_dir):
         shutil.copytree(example_dir, upgrade_dir, ignore=shutil.ignore_patterns('outputs'))
         upgrade_dir_inputs = os.path.join(upgrade_dir, 'inputs')
         upgrade_dir_outputs = os.path.join(upgrade_dir, 'outputs')
-        upgrade_inputs(upgrade_dir_inputs)
+        switch_model.upgrade.manager.set_verbose(False)
         try:
             # Custom python modules may be in the example's working directory
+            upgrade_inputs(upgrade_dir_inputs)
             sys.path.append(upgrade_dir)
             switch_model.solve.main([
                 '--inputs-dir', upgrade_dir_inputs,
